(PC-6695) Move functions in utils file
